fix(review): split context def to fix only-export-components
