+++ src/components/NotificationToast.test.tsx
@@ -2,7 +2,8 @@ import { describe, expect, it } from 'vitest'
 import { render, screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
 import { NotificationToast } from './NotificationToast'
-import { NotificationContext, NotificationProvider } from '../contexts/NotificationContext'
+import { NotificationContext } from '../contexts/notification-context'
+import { NotificationProvider } from '../contexts/NotificationContext'
 import { useContext } from 'react'
 
 function TestWrapper({ children }: { children: React.ReactNode }) {
