+++ src/contexts/notification-context.ts
@@ -1,0 +1,29 @@
+import { createContext } from 'react'
+
+type NotificationType = 'success' | 'error' | 'warning' | 'info'
+
+interface Notification {
+  id: string
+  type: NotificationType
+  message: string
+  duration?: number
+}
+
+interface NotificationContextType {
+  notifications: Notification[]
+  addNotification: (type: NotificationType, message: string, duration?: number) => void
+  removeNotification: (id: string) => void
+  success: (message: string, duration?: number) => void
+  error: (message: string, duration?: number) => void
+  warning: (message: string, duration?: number) => void
+  info: (message: string, duration?: number) => void
+}
+
+const NotificationContext = createContext<NotificationContextType | undefined>(undefined)
+
+export {
+  type NotificationType,
+  type Notification,
+  type NotificationContextType,
+  NotificationContext,
+}
