+++ src/hooks/useNotification.ts
@@ -1,5 +1,6 @@
 import { useContext } from 'react'
-import { NotificationContext } from '../contexts/NotificationContext'
+
+import { NotificationContext } from '../contexts/notification-context'
 
 export function useNotification() {
   const context = useContext(NotificationContext)
